==> 02-standardize_plates.Rmd <==
---
title: "02-standardize_plates"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# packages

```{r,echo=TRUE,warning=FALSE,message=FALSE}
library(tidyverse)
theme_set(theme_bw())
```

# import plates in long-format

```{r}
data4 <- read_rds("data/multiplate-long.rds")
```

```{r}
data4 %>% rmarkdown::paged_table()
```


# create standardization dataframe

```{r}
std_reference <- 
  tribble(
    ~dilution,~reciprocal,~factor,~abunits,
    "STD 1/50",50,"102400/50",NA_real_,
    "STD 1/100",100,"102400/100",NA_real_,
    "STD 1/200",200,"102400/200",NA_real_,
    "STD 1/400",400,"102400/400",NA_real_,
    "STD 1/800",800,"102400/800",NA_real_,
    "STD 1/1600",1600,"102400/1600",NA_real_,
    "STD 1/3200",3200,"102400/3200",NA_real_,
    "STD 1/6400",6400,"102400/6400",NA_real_,
    "STD 1/12800",12800,"102400/12800",NA_real_,
    "STD 1/25600",25600,"102400/25600",NA_real_,
    "STD 1/51200",51200,"102400/51200",NA_real_,
    "STD 1/102400",102400,"102400/102400",NA_real_,
  ) %>% 
  mutate(abunits=max(reciprocal)/reciprocal)

std_reference
```

# build density ~ abunit relationship

## isolate standard curves

```{r}
data5 <- data4 %>% 
  filter(str_detect(codigo,"STD")) %>% 
  left_join(std_reference,by = c("codigo"="dilution"))
```

```{r}
data5 %>% rmarkdown::paged_table()
```

```{r}
data6 <- data4 %>% 
  filter(str_detect(codigo,"ctrl"))
```

## descriptive

```{r}
hospital_names <- list(
  'igg'="IgG",
  'igg-avidez'="IgG avidity"
)

hospital_labeller <- function(variable,value){
  return(hospital_names[value])
}
```


```{r,fig.height=10,fig.width=10}
figure01 <- 
  data5 %>% 
  ggplot(aes(x = abunits,y = densidad)) +
  geom_point() +
  # geom_smooth() +
  geom_hline(
    aes(yintercept=densidad, color = codigo),
    data = 
      data6 %>% 
      mutate(
        codigo = 
          case_when(
            codigo == "ctrl_blanco" ~ "Blank",
            codigo == "ctrl_negativo" ~ "Negative",
            codigo == "ctrl_positivo" ~ "Positive"
          )
      )
  ) +
  scale_x_log10() +
  facet_wrap(~plate+experimento,labeller = hospital_labeller) +
  labs(
    x = "Arbitrary Antibody Units",
    y = "Optical Density",
    color = "Controls"
  )

figure01

ggsave(
  filename = "figure/02-figure01.png",
  plot = figure01,
  width = 10,height = 10
)

ggsave(
  filename = "figure/02-figure01.tiff",
  plot = figure01,
  width = 10,height = 10
)
```

## log-logistic model

- [pending]

```{r}

```

# quality control of replicates

## replicates per plate

- how many sample replicates are in each plate?

```{r,message=FALSE}
data4 %>% 
  filter(!str_detect(codigo,"STD|ctrl")) %>% 
  count(plate,experimento,codigo,sort = T) %>% 
  count(n)

data4 %>% 
  filter(!str_detect(codigo,"STD|ctrl")) %>% 
  count(plate,experimento,codigo,sort = T) %>% 
  # epihelper::print_inf() %>% 
  identity()
```

## replicates all plate

- how many sample replicates are in all plates?

```{r,message=FALSE}
data4 %>% 
  filter(!str_detect(codigo,"STD|ctrl")) %>% 
  count(codigo,sort = T) %>% 
  count(n)

data4 %>% 
  filter(!str_detect(codigo,"STD|ctrl")) %>% 
  count(codigo,sort = T)
```

## replicate type

```{r}
data4 %>% 
  filter(str_detect(codigo,"ctrl")) %>% 
  count(plate,experimento,codigo) %>% 
  identity()
  # group_by(plate,experimento,codigo) %>% 
  # skimr::skim(n)
```


## replicate variability

actions:

- remove standard curve rows
- summarized samples and control wells to mean, sd, and cv

```{r}
data7 <- data4 %>% 
  # filter(!str_detect(codigo,"STD|ctrl")) %>% 
  filter(!str_detect(codigo,"STD")) %>% 
  # here all replicates are summarized
  # replicates from samples and controls
  group_by(plate,experimento,codigo) %>% 
  summarise(densidad_replicate = n(),
            densidad_mean=mean(densidad),
            densidad_sd=sd(densidad)) %>% 
  ungroup() %>% 
  mutate(densidad_cv=100*densidad_sd/densidad_mean) %>% 
  # epihelper::print_inf() %>% 
  identity()

# data7
```

```{r}
data7 %>% rmarkdown::paged_table()
```


### identify outliers

```{r}
data8 <- data7 %>% 
  filter(densidad_mean>900)

data8
```

```{r}
data8 %>% 
  select(plate,codigo)
data4 %>% 
  filter(magrittr::is_in(
    x = plate,
    table = data8 %>% pull(plate))) %>% 
  filter(magrittr::is_in(
    x = codigo,
    table = data8 %>% pull(codigo)))
```


### distribution table

#### con outlier

```{r}
data7 %>% 
  select(starts_with("densidad")) %>% 
  skimr::skim() %>% 
  skimr::yank(skim_type = "numeric")
```

#### sin outlier

```{r}
data7 %>% 
  filter(densidad_mean<900) %>% 
  select(starts_with("densidad")) %>% 
  skimr::skim() %>% 
  skimr::yank(skim_type = "numeric")
```

### distribution plot

#### con outlier

```{r,fig.height=10,fig.width=10}
figure02 <- 
  data7 %>% 
  ggplot(aes(x = densidad_mean,y = densidad_cv)) +
  geom_point() +
  geom_hline(aes(yintercept=20),
             linetype="dashed",size=0.3) +
  geom_vline(aes(xintercept=0.25),
             linetype="dashed",size=0.3) +
  coord_cartesian(ylim = c(0,100)) +
  facet_wrap(~plate+experimento, labeller = hospital_labeller) +
  labs(
    x = "Mean Optical Density",
    y = "Coefficient of Variation of replicates"
  )

figure02

ggsave(
  filename = "figure/02-figure02.png",
  plot = figure02,
  width = 10,height = 10
)

ggsave(
  filename = "figure/02-figure02.tiff",
  plot = figure02,
  width = 10,height = 10
)
```

#### sin outlier

```{r,fig.height=10,fig.width=10}
data7 %>% 
  filter(densidad_mean<900) %>% 
  ggplot(aes(x = densidad_mean,y = densidad_cv)) +
  geom_point() +
  geom_hline(aes(yintercept=20),
             linetype="dashed",size=0.3) +
  geom_vline(aes(xintercept=0.25),
             linetype="dashed",size=0.3) +
  coord_cartesian(ylim = c(0,100)) +
  facet_wrap(~plate+experimento)
```

# view plate shape

```{r,eval=FALSE}
plater::view_plate(
  data = data4 %>% filter(plate=="placa_001"), 
  well_ids_column = "pozo", 
  columns_to_display = c("codigo","densidad")
)

plater::view_plate(
  data = data4 %>% filter(plate=="placa_002"), 
  well_ids_column = "pozo", 
  columns_to_display = c("codigo","densidad")
)
```

# output data sets

```{r}
data9 <- data4 %>% 
  filter(!str_detect(codigo,"STD")) #%>% 
  # filter(str_detect(codigo,"ctrl"))
```

```{r}
# data4 # all 
# only std + density ~ abunits
data5 %>% write_rds("data/02-data-standard_curve.rds") 
# data6 # only ctrl
data6 %>% write_rds("data/02-data-only_control.rds") 
# samples and ctrl with stat measurements per code
data7 %>% write_rds("data/02-data-ab_mean.rds") 
# data8 # outliers if any
# data9 # only samples and ctrl
```

# to do list

- [ ] create `std_reference` from data! flexible to any new situation
