==> 04-purrr-estimate_abunits.Rmd <==
---
title: "04-purrr-estimate_abunits"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# packages

```{r,echo=TRUE,warning=FALSE,message=FALSE}
library(tidyverse)
library(drc)
library(conflicted)
conflicts_prefer(dplyr::select)
conflicts_prefer(dplyr::filter)
```

```{r}
options(scipen=10000)
```

# import input data

```{r}
data_stdc <- read_rds("data/02-data-standard_curve.rds")
data_ctrl <- read_rds("data/02-data-only_control.rds")
data_abmean <- read_rds("data/02-data-ab_mean.rds") 
```

```{r}
data_stdc %>% rmarkdown::paged_table()
```

```{r}
data_abmean %>% rmarkdown::paged_table()
```

# fit 4pll model

## dose response model

```{r,eval=FALSE}
# one plate experiment
wb.model <- 
  drm(formula = densidad ~ abunits, 
    curveid = plate,
    data= data_drm_stdc,
    fct = LL.4(names = c("b", "c", "d", "e")))
#' "b" -> "Slope", 
#' "c" -> "Lower Limit", 
#' "d" -> "Upper Limit", 
#' "e" -> "ED50"

wb.model %>% broom::tidy()
wb.model %>% broom::glance()
# wb.model %>% pluck("data")
```

```{r,warning=FALSE}
# multiple plate experiment
data_drm_models <- 
  data_stdc %>%
  # data_drm_stdc %>%
  group_by(plate,experimento) %>% 
  nest() %>% 
  ungroup() %>% 
  mutate(fit = 
           map(.x = data, 
               .f = ~drm(formula = densidad ~ abunits, 
                         # curveid = plate,
                         data= .x,
                         fct = LL.4(names = c("b", "c", "d", "e")))
               )
         ) %>% 
  # mutate(fit_tidy = map(.x = fit,.f = broom::tidy)) %>% 
  # mutate(fit_glance = map(.x = fit,.f = broom::glance)) %>% 
  identity()

data_drm_models
# data_drm_models %>% 
#   unnest(cols = fit_tidy)
# data_drm_models %>% 
#   unnest(cols = fit_glance)
```

## box-cox transformation

BOX-COX TRANSFORMATION against RESIDUAL heterogeneity

```{r,eval=FALSE}
# only one plate

wb.model.BX <- 
  boxcox(object = wb.model, 
         # main=expression("Optimal " ~ lambda ~ " with confidence intervals"), 
         plotit = FALSE)

wb.model.BX %>% broom::tidy()
```

```{r,warning=FALSE}
# for multiple plates

data_output <- NULL

for (i in 1:nrow(data_drm_models)) {
  
  test01 <- data_drm_models %>% 
    # dplyr::select(plate,experimento,data,fit) %>% 
    slice(i)
  
  .x <- test01 %>%
    pull(data) %>%
    pluck(1)
  
  test02 <- test01 %>%
    pull(fit) %>%
    pluck(1) %>%
    # class()
    boxcox(object = .,plotit = FALSE)
  
  test03 <- tibble(fit_boxcox = list(test02))
  
  test04 <- test01 %>% add_column(test03)
  
  data_output <- union_all(data_output,test04)
  
}

data_output
```

## tidy model outputs

```{r}
data_drm_models_tidy <- 
  data_output %>% 
  mutate(fit_tidy = map(.x = fit_boxcox,.f = broom::tidy)) %>%
  mutate(fit_glance = map(.x = fit_boxcox,.f = broom::glance))

data_drm_models_tidy %>%
  unnest(cols = fit_tidy)
data_drm_models_tidy %>%
  unnest(cols = fit_glance)
```

# predict the standard curve

## prediction and confidence interval

```{r}
data_new <- expand.grid(exp(seq(log(0.1),log(2048),length=100)))
```

```{r,eval=FALSE}
# only one plate

# predictions and confidence intervals
pdm <- predict(object = wb.model.BX, newdata = data_new, interval = "confidence")

## join data
data_new_pred <- data_new %>% 
  rownames_to_column() %>% 
  left_join(
    pdm %>% 
      as_tibble() %>% 
      rownames_to_column()) %>% 
  mutate(plate=unique(data_stdc$plate)[j]) %>% 
  dplyr::rename(resp=Var1,p=Prediction,pmin=Lower,pmax=Upper) %>% 
  dplyr::select(plate,everything())

```


```{r,warning=FALSE}
# multiple plates

data_output_02 <- NULL

for (i in 1:nrow(data_drm_models_tidy)) { #i=1
  
  test11 <- data_drm_models_tidy %>% 
    slice(i)
  
  test12 <- predict(object = test11 %>% 
                      pull(fit_boxcox) %>% 
                      pluck(1), 
                    newdata = data_new, 
                    interval = "confidence")
  
  test13 <- data_new %>% 
    rownames_to_column() %>% 
    left_join(
      test12 %>% 
        as_tibble() %>% 
        rownames_to_column()) %>% 
    # mutate(plate=unique(data_stdc$plate)[j]) %>% 
    dplyr::rename(resp=Var1,p=Prediction,pmin=Lower,pmax=Upper) %>% 
    dplyr::select(everything())
  
  test14 <- tibble(prediction = list(test13))
  
  test15 <- test11 %>% add_column(test14)
  
  data_output_02 <- union_all(data_output_02,test15)
  
}

data_output_02

```

## result

```{r}
hospital_names <- list(
  'igg'="IgG",
  'igg-avidez'="IgG avidity"
)

hospital_labeller <- function(variable,value){
  return(hospital_names[value])
}
```

```{r,fig.height=10,fig.width=10}
data_output_02_pred <- data_output_02 %>%
  dplyr::select(plate,experimento,prediction) %>% 
  unnest(cols = prediction)

figure01 <- 
  ggplot() +
  geom_hline(
    mapping = aes(yintercept=densidad, color = codigo),
    data = 
      data_ctrl %>% 
      mutate(
        codigo = 
          case_when(
            codigo == "ctrl_blanco" ~ "Blank",
            codigo == "ctrl_negativo" ~ "Negative",
            codigo == "ctrl_positivo" ~ "Positive"
          )
      )
  ) +
  geom_ribbon(data = data_output_02_pred,
              mapping = aes(ymin = pmin, ymax = pmax, 
                            x = resp), fill = "grey70") +
  geom_line(data = data_output_02_pred,
            mapping = aes(x = resp,y = p)) +
  geom_point(data = data_stdc,
             mapping = aes(x = abunits,y = densidad)) +
  scale_x_log10() +
  facet_wrap(~plate+experimento, labeller = hospital_labeller) +
  labs(
    x = "Arbitrary Antibody Units",
    y = "Optical Density", 
    color = "Controls"
    # x = "Unidades de Anticuerpos",
    # y = "Densidad óptica"
  )

figure01

ggsave(
  filename = "figure/04-figure01.png",
  plot = figure01,
  width = 10,height = 10
)

ggsave(
  filename = "figure/04-figure01.tiff",
  plot = figure01,
  width = 10,height = 10
)
```

# estimate abunits

UNK AB.UNITS ESTIMATION by INVERSE REGRESSION [pending]

## estimates and confidence intervals

```{r,eval=FALSE}
# only one plate

data_ed_abunit <-
  ED(object = wb.model.BX,
     respLev = data_ed_abmean %>% pull(densidad_mean),
     #j #wb_MEAN[1:n,5],
     type = "absolute",
     interval = "delta",
     #clevel = "Pfal",
     display = FALSE)

data_abmean_abunit <-
  data_ed_abmean %>%
  rownames_to_column() %>%
  left_join(
    data_ed_abunit %>%
      as_tibble() %>%
      rownames_to_column() %>%
      janitor::clean_names()
  )
```

```{r,warning=FALSE}
# multiple plates

data_abmean_nested <- data_abmean %>% 
  group_by(plate,experimento) %>% 
  nest() %>% 
  ungroup()

data_output_03 <- NULL

for (i in 1:nrow(data_output_02)) {#i=1
  
  test21 <- data_output_02 %>% 
    slice(i)
  
  test22 <- data_abmean_nested %>% 
    slice(i)
  
  test22_abmean <- test22 %>% 
         pull(data) %>% 
         pluck(1)
  
  test23 <-
    ED(object = test21 %>% 
         pull(fit_boxcox) %>% 
         pluck(1),
       respLev = test22_abmean %>% 
         pull(densidad_mean),
       #j #wb_MEAN[1:n,5],
       type = "absolute",
       interval = "delta",
       #clevel = "Pfal",
       display = FALSE)
  
  # create rowname column name
  # useful only for joining two datasets
  test24 <- test22_abmean %>%
    rownames_to_column() %>%
    left_join(
      test23 %>%
        as_tibble() %>%
        rownames_to_column() %>%
        janitor::clean_names()
    ) %>% 
    dplyr::select(-rowname)
  
  test25 <- tibble(estimation = list(test24))
  
  test26 <- test21 %>% add_column(test25)
  
  data_output_03 <- union_all(data_output_03,test26)
  
}

data_output_03
```

## result

```{r}
data_output_03_abunit <- data_output_03 %>%
  dplyr::select(plate,experimento,estimation) %>% 
  unnest(cols = estimation)
```

```{r}
data_output_03_abunit %>% rmarkdown::paged_table()
```


```{r,fig.height=10,fig.width=10}
figure02 <- 
  ggplot() +
  geom_hline(
    mapping = aes(yintercept=densidad, color = codigo),
    data = 
      data_ctrl %>% 
      mutate(
        codigo = 
          case_when(
            codigo == "ctrl_blanco" ~ "Blank",
            codigo == "ctrl_negativo" ~ "Negative",
            codigo == "ctrl_positivo" ~ "Positive"
          )
      )
  ) +
  geom_ribbon(data = data_output_02_pred,
              mapping = aes(ymin = pmin, ymax = pmax, 
                            x = resp), fill = "grey70") +
  geom_line(data = data_output_02_pred,
            mapping = aes(x = resp,y = p)) +
  geom_point(mapping = aes(x = estimate,y = densidad_mean),
             data = data_output_03_abunit) +
  geom_errorbar(mapping = aes(x = estimate,y = densidad_mean,
                    xmin = lower, xmax = upper),
                data = data_output_03_abunit) +
  scale_x_log10() +
  facet_wrap(~plate+experimento, labeller = hospital_labeller) +
  labs(
    x = "Arbitrary Antibody Units",
    y = "Optical Density", 
    color = "Controls"
    # x = "Unidades de Anticuerpos",
    # y = "Densidad óptica"
  )

figure02

ggsave(
  filename = "figure/04-figure02.png",
  plot = figure02,
  width = 10,height = 10
)

ggsave(
  filename = "figure/04-figure02.tiff",
  plot = figure02,
  width = 10,height = 10
)
```

# outputs

```{r}
data_output_03 %>% 
  write_rds("data/03-data-ab_units-complete.rds")

data_output_03_abunit %>% 
  writexl::write_xlsx("data/03-data-ab_units-only.xlsx")
```

# iterative modeling

- [x] overlap standard curve plate points with model
- [x] replicate the ggplot2 aesthetic from reference
- [x] use purrr to model to all stratas
- [x] identify the reference code for classification!
- review https://rstats4ag.org/dose-response-curves.html
- review https://bioassay.dk/

# questions

- [ ] how to treat NaN's in drc output?
