==> 05-serological_classification.Rmd <==
---
title: "05-serological_classification"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# packages

```{r,echo=TRUE,warning=FALSE,message=FALSE}
library(tidyverse)
library(mixtools)
library(patchwork)
library(conflicted)
conflicts_prefer(dplyr::select)
conflicts_prefer(dplyr::filter)
```

```{r}
options(scipen=10000)
```

# import input data

```{r}
data_output_03 <- read_rds("data/03-data-ab_units-complete.rds")
```

## extract antibody units estimates

```{r}
data_output_03_abunit <- data_output_03 %>%
  dplyr::select(plate,experimento,estimation) %>% 
  unnest(cols = estimation) %>% 
  
  # extra code to create unique id
  mutate(unique_id = paste0(#experimento,"-",
                            plate,"--",codigo))
```

```{r,eval=FALSE,echo=FALSE}
# data_output_03
# data_output_03_abunit %>% 
#   filter(experimento!="igg_______") %>% 
#   distinct(plate,codigo)
```

```{r}
data_output_03_abunit %>% rmarkdown::paged_table()
```

## reminder

observations are unique when stratified by `experimento` and combined by `plate`, `codigo`:

```{r}
data_output_03_abunit %>% 
  filter(str_detect(codigo,"ctrl_")) %>% 
  filter(magrittr::is_in(plate,c("placa_001","placa_002"))) %>% 
  select(experimento,unique_id) %>% 
  arrange(experimento)
```

# explore data

```{r}
hospital_names <- list(
  'igg'="IgG",
  'igg-avidez'="IgG avidity"
)

hospital_labeller <- function(variable,value){
  return(hospital_names[value])
}
```

## antibody distribution in sample

```{r,fig.height=6,fig.width=6,warning=FALSE}
fig01_a <- data_output_03_abunit %>% 
  ggplot(aes(x = estimate)) +
  geom_density(alpha=.5, 
               position = "identity") +
  geom_histogram(aes(y = after_stat(density)),
                 bins = 60,
                 alpha=.5, 
                 position = "identity") +
  facet_grid(~experimento,scales = "free", labeller = hospital_labeller) +
  #theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  labs(
    title="Linear distribution",
    x = "Arbitrary Antibody Units",
    y = "Density"
  )

fig01_b <- data_output_03_abunit %>% 
  ggplot(aes(sample = estimate)) +
  geom_qq(alpha=.2) +
  geom_qq_line(line.p = c(0.25, 0.75)) +
  facet_grid(~experimento,scales = "free", labeller = hospital_labeller) +
  labs(title="Gaussian quantile-quantile plot") +
  coord_cartesian(ylim = c(0,2000)) +
  labs(
    x = "Theoretical quantiles",
    y = "Sample quantiles"
  )

figure01 <- fig01_a / fig01_b

figure01

ggsave(
  filename = "figure/05-figure01.png",
  plot = figure01,
  width = 6,height = 6
)

ggsave(
  filename = "figure/05-figure01.tiff",
  plot = figure01,
  width = 6,height = 6
)
```

## identify missing values

```{r}
data_output_03_abunit %>% 
  select(experimento,unique_id,estimate) %>% 
  naniar::vis_miss()
```


# fit a gaussian mixture model

Use `mixtools` package following the blog post about [Using Mixture Models for Clustering](https://tinyheero.github.io/2015/10/13/mixture-model.html)

Use it for two `experimento` and at least two `k-component` variations `k=2` and `k=3`

## for one strata

```{r}
data_output_03_abunit %>% 
  select(experimento,estimate) %>% 
  filter(!is.na(estimate)) %>% 
  filter(experimento=="igg") %>% 
  pull(estimate) %>% 
  normalmixEM(k=2) %>% 
  class()
```

## for all stratas and all components

```{r}
model_output_01 <- data_output_03_abunit %>% 
  # filter(experimento=="igg") %>% 
  select(experimento,unique_id,densidad_replicate,densidad_mean,densidad_sd,estimate) %>% 
  # ojo: lost of sample of density values with missing estimates 
  filter(!is.na(estimate)) %>% 
  group_by(experimento) %>% 
  nest() %>% 
  ungroup() %>%
  mutate(estimate_strata = map(data,pull,estimate)) %>% 
  # expand the grid by the combination of k components
  expand_grid(component_n = 2:5) %>%
  mutate(mixture_model = 
           pmap(.l = 
                  select(
                    .,
                    x = estimate_strata,
                    k = component_n
                  ),
                .f = normalmixEM)) %>% 
  # extract the log likelihood
  mutate(loglikelihood = map_dbl(mixture_model,pluck,"loglik")) %>% 
  # compute the aic
  mutate(aic=2*component_n-2*loglikelihood) %>% 
  group_by(experimento) %>% 
  mutate(aic_difference = aic - dplyr::lag(aic)) %>% 
  ungroup()

# model_output_01
```

## criteria for model selection

criteria to choose the number of `k-compartments` for a distribution:

- not based only on AIC criteria, but also in visual inspection of the distribution
- minimize the number of undetermined samples (`s0`)
- closer resemblance to the expected number of positives

## summary of model performance

```{r}
model_output_01 %>% 
  select(experimento,component_n,loglikelihood,aic,aic_difference)
```


## visualize mixture model performance

### function to plot mixture components

```{r}
#' Plot a Mixture Component
#' 
#' @param x Input data
#' @param mu Mean of component
#' @param sigma Standard deviation of component
#' @param lam Mixture weight of component
plot_mix_comps <- function(x, mu, sigma, lam) {
  lam * dnorm(x, mu, sigma)
}

stat_mixture <- function(model_name, k_component, colour) {
  stat_function(geom = "line", fun = plot_mix_comps,
                args = list(model_name$mu[k_component], 
                            model_name$sigma[k_component], 
                            lam = model_name$lambda[k_component]),
                colour = colour, lwd = .25)
}
```

components distributions equivalent to multiple latent seropositive states:

| component|     color|     state|
|:---------|:---------|:---------|
|    comp.1|       red|        s-|
|    comp.2|      blue|        s+|
|    comp.3|     green|       s++|

### experiment 1

```{r}
model_igg_k2 <- model_output_01 %>% 
  filter(experimento == "igg",
         component_n == 2) %>% 
  pull(mixture_model) %>% 
  pluck(1)

# manual_colors <- c("seronegative" = "red",
#                    "seropositive" = "blue")


fig02a <- data.frame(x = model_igg_k2$x) %>%
  mutate(experiment = "igg") %>% 
  ggplot() +
  geom_histogram(aes(x, after_stat(density)), 
                 bins = 30,
                 alpha=.5,
                 position = "identity") +
  facet_wrap(~experiment,labeller = hospital_labeller) +
  stat_mixture(model_name = model_igg_k2,
               k_component = 1,
               colour = "red") +
  stat_mixture(model_name = model_igg_k2,
               k_component = 2,
               colour = "blue") +
  labs(
    # title="Linear distribution",
    x = "Arbitrary Antibody Units",
    y = "Density",
    caption = "Two distributions:\n- Red: seronegative\n- Blue: seropositive"
  ) +
  # scale_colour_manual(values = manual_colors)
  theme(plot.caption = element_text(hjust = 0)) # set the left align here
```

```{r}
model_igg_k3 <- model_output_01 %>% 
  filter(experimento == "igg",
         component_n == 3) %>% 
  pull(mixture_model) %>% 
  pluck(1)

fig02b <- data.frame(x = model_igg_k3$x) %>%
  mutate(experiment = "igg") %>% 
  ggplot() +
  geom_histogram(aes(x, after_stat(density)), 
                 bins = 30,
                 alpha=.5,
                 position = "identity") +
  facet_wrap(~experiment, labeller = hospital_labeller) +
  stat_mixture(model_name = model_igg_k3,k_component = 1,colour = "red") +
  stat_mixture(model_name = model_igg_k3,k_component = 2,colour = "blue") +
  stat_mixture(model_name = model_igg_k3,k_component = 3,colour = "green") +
  labs(
    # title="Linear distribution",
    x = "Arbitrary Antibody Units",
    y = "Density",
    caption = "red : seronegative\nblue : seropositive\ngreen : boosting upon recurrent infection"
  ) +
  theme(plot.caption = element_text(hjust = 0)) # set the left align here
```

```{r,fig.height=3,fig.width=7}
fig02a + fig02b
```

### experiment 2

```{r}
model_avi_k2 <- model_output_01 %>% 
  filter(experimento == "igg-avidez",
         component_n == 2) %>% 
  pull(mixture_model) %>% 
  pluck(1)

fig03a <- data.frame(x = model_avi_k2$x) %>%
  mutate(experiment = "igg-avidez") %>% 
  ggplot() +
  geom_histogram(aes(x, after_stat(density)), 
                 bins = 30,
                 alpha=.5,
                 position = "identity") +
  facet_wrap(~experiment, labeller = hospital_labeller) +
  stat_mixture(model_name = model_avi_k2,k_component = 1,colour = "red") +
  stat_mixture(model_name = model_avi_k2,k_component = 2,colour = "blue") +
  labs(
    # title="Linear distribution",
    x = "Arbitrary Antibody Units",
    y = "Density",
    caption = "Two distributions:\n- Red: seronegative\n- Blue: seropositive"
  ) +
  theme(plot.caption = element_text(hjust = 0)) # set the left align here
```

```{r}
model_avi_k3 <- model_output_01 %>% 
  filter(experimento == "igg-avidez",
         component_n == 3) %>% 
  pull(mixture_model) %>% 
  pluck(1)

fig03b <- data.frame(x = model_avi_k3$x) %>%
  mutate(experiment = "igg-avidez") %>% 
  ggplot() +
  geom_histogram(aes(x, after_stat(density)), 
                 bins = 30,
                 alpha=.5,
                 position = "identity") +
  facet_wrap(~experiment, labeller = hospital_labeller) +
  stat_mixture(model_name = model_avi_k3,k_component = 1,colour = "red") +
  stat_mixture(model_name = model_avi_k3,k_component = 2,colour = "blue") +
  stat_mixture(model_name = model_avi_k3,k_component = 3,colour = "green") +
  labs(
    # title="Linear distribution",
    x = "Arbitrary Antibody Units",
    y = "Density",
    caption = "red : seronegative\nblue : seropositive\ngreen : boosting upon recurrent infection"
  ) +
  theme(plot.caption = element_text(hjust = 0)) # set the left align here
```

```{r,fig.height=3,fig.width=7}
fig03a + fig03b
```

### choice

```{r,fig.height=3,fig.width=7}
figure02 <- fig02a + fig03a

figure02

ggsave(
  filename = "figure/05-figure02-comp2.png",
  plot = figure02,
  height = 3.5, width = 7
)

ggsave(
  filename = "figure/05-figure02-comp2.tiff",
  plot = figure02,
  height = 3.5, width = 7
)
```

# classify seropositivity

## define probability threshold

```{r}
class_probability <- 0.90
```

## classify using threshhold

```{r}
data_output_04 <- model_output_01 %>% 
  filter(magrittr::is_in(component_n,c(2,3))) %>% 
  select(experimento,data,component_n,mixture_model) %>% 
  mutate(mixture_components = map(
    .x = mixture_model,
    .f = pluck, "posterior")) %>%
  mutate(mixture_components = map(
    .x = mixture_components, 
    .f = as_tibble)) %>% 
  select(-mixture_model) %>% 
  unnest(cols = c(data,component_n,mixture_components)) %>% 
  # for a clear number output
  # mutate(across(.cols = where(is.numeric),.fns = round, 8)) %>% 
  mutate(across(.cols = where(is.numeric),.fns = \(x) round(x,8))) %>% 
  # rmarkdown::paged_table()
  # # quality control
  # count(experimento,component_n)
  # rowwise to securily sum numbers removing missing values
  rowwise() %>% 
  mutate(comp.23 = sum(c_across(cols = c(comp.2,comp.3)),na.rm = T)) %>% 
  ungroup() %>% 
  # k-labeling
  # two alternatives open to identify a superinfection latent class
  mutate(class = case_when(
    comp.1 > class_probability ~ "s-",
    comp.2 > class_probability ~ "s+",
    comp.3 > class_probability ~ "s++",
    TRUE ~ "s0"
  )) %>% 
  mutate(class_sum = case_when(
    comp.1 > class_probability ~ "s-",
    comp.23 > class_probability ~ "s+",
    TRUE ~ "s0"
  )) %>% 
  mutate(class = fct_relevel(class, "s-","s0","s+","s++")) %>% 
  mutate(class_sum = fct_relevel(class_sum, "s-","s0","s+")) %>% 
  # separate the unique id in case useful for user
  separate(col = unique_id,into = c("plate","codigo"),sep = "--",remove = F) %>% 
  # send unique to the end
  relocate(unique_id, .after = last_col())
```

```{r,eval=FALSE}
data_output_04 #%>% rmarkdown::paged_table()
```

```{r,eval=FALSE,echo=FALSE}
# quality control
# all prob must sum 1
data_output_04 %>% 
  select(starts_with("comp.")) %>% 
  rowwise() %>% 
  mutate(sum = sum(c_across(cols = -comp.23),na.rm = T)) %>% 
  ungroup() %>% 
  select(sum) %>% 
  summary()
```


## summary of classification

```{r}
data_output_04 %>% 
  count(experimento,component_n,class)

data_output_04 %>% 
  count(experimento,component_n,class_sum)
```

## visualize classification

### by distribution

```{r,fig.height=5,fig.width=8}
figure04_1 <- 
  data_output_04 %>% 
  ggplot(aes(x = estimate, 
             fill = class, 
             after_stat(density))) +
  geom_histogram(binwidth = 50,
                 alpha=.5,
                 position = "identity") +
  facet_wrap(~experimento+component_n,scales = "free",labeller = hospital_labeller) +
  labs(
    # title="Linear distribution",
    x = "Arbitrary Antibody Units",
    y = "Density",
    fill = "Classification",
    caption = "s- : seronegative\ns0 : undetermined\ns+ : seropositive\ns++ : boosting upon recurrent infection"
  ) +
  theme(plot.caption = element_text(hjust = 0)) # set the left align here

figure04_1

# ggsave(
#   filename = "figure/05-figure04_1.png",
#   plot = figure04_1,
#   height = 5.5, width = 8
# )
# 
# ggsave(
#   filename = "figure/05-figure04_1.tiff",
#   plot = figure04_1,
#   height = 5.5, width = 8
# )
```

```{r,fig.height=5,fig.width=8}
figure04_2 <- 
  data_output_04 %>% 
  ggplot(aes(x = estimate,
             fill = class_sum, 
             after_stat(density))) +
  geom_histogram(binwidth = 50,
                 alpha=.5,
                 position = "identity") +
  facet_wrap(~experimento+component_n,scales = "free",labeller = hospital_labeller) +
  labs(
    # title="Linear distribution",
    x = "Arbitrary Antibody Units",
    y = "Density",
    fill = "Classification",
    caption = "s- : seronegative\ns0 : undetermined\ns+ : seropositive"
  ) +
  theme(plot.caption = element_text(hjust = 0)) # set the left align here

figure04_2

# ggsave(
#   filename = "figure/05-figure04_2.png",
#   plot = figure04_2,
#   height = 5.5, width = 8
# )
# 
# ggsave(
#   filename = "figure/05-figure04_2.tiff",
#   plot = figure04_2,
#   height = 5.5, width = 8
# )
```


### by frecuency

```{r,fig.height=5.5,fig.width=6}
figure05_1 <- 
  data_output_04 %>% 
  ggplot(aes(x = class)) +
  geom_bar() +
  facet_wrap(~experimento+component_n,labeller = hospital_labeller) +
  labs(
    # title="Linear distribution",
    x = "Serological Classification",
    y = "Frequency",
    caption = "s- : seronegative\ns0 : undetermined\ns+ : seropositive\ns++ : boosting upon recurrent infection"
  ) +
  theme(plot.caption = element_text(hjust = 0)) # set the left align here
```

```{r,fig.height=5.5,fig.width=6}
figure05_2 <- 
  data_output_04 %>% 
  ggplot(aes(x = class_sum)) +
  geom_bar() +
  facet_wrap(~experimento+component_n, labeller = hospital_labeller) +
  labs(
    # title="Linear distribution",
    x = "Serological Classification",
    y = "Frequency",
    caption = "s- : seronegative\ns0 : undetermined\ns+ : seropositive"
  ) +
  theme(plot.caption = element_text(hjust = 0)) # set the left align here
```

#### choice

```{r,fig.height=5.5,fig.width=6}
data_output_04 %>% 
  count(component_n)

figure05 <- 
  data_output_04 %>%
  filter(component_n == 2) %>% 
  ggplot(aes(x = class_sum)) +
  geom_bar() +
  facet_wrap(~experimento, labeller = hospital_labeller
             # +component_n
             ) +
  labs(
    # title="Linear distribution",
    x = "Serological Classification",
    y = "Frequency",
    caption = "s- : seronegative\ns0 : undetermined\ns+ : seropositive"
  ) +
  theme(plot.caption = element_text(hjust = 0)) # set the left align here

figure05

ggsave(
  filename = "figure/05-figure05-comp2.png",
  plot = figure05,
  height = 3.5, width = 6
)

ggsave(
  filename = "figure/05-figure05-comp2.tiff",
  plot = figure05,
  height = 3.5, width = 6
)
```


### by classification probability

```{r,fig.height=5.5,fig.width=7}
figure06_1 <- 
  data_output_04 %>% 
  ggplot() +
  geom_line(aes(x = estimate,y = comp.1), colour="red", lwd = 1.5) +
  geom_line(aes(x = estimate,y = comp.2), colour="blue", lwd = 1.5) +
  geom_line(aes(x = estimate,y = comp.3), colour="green", lwd = 1.5) +
  geom_hline(yintercept = class_probability, col = "black") +
  facet_wrap(~experimento+component_n,scales = "free",labeller = hospital_labeller) +
  labs(
    # title="Linear distribution",
    x = "Arbitrary Antibody Units",
    y = "Classification Probability",
    caption = "red : seronegative\nblue : seropositive\ngreen : boosting upon recurrent infection\nblack : 90% classification probability"
  ) +
  theme(plot.caption = element_text(hjust = 0)) # set the left align here

figure06_1

# ggsave(
#   filename = "figure/05-figure06_1.png",
#   plot = figure06_1,
#   height = 5.5, width = 7
# )
# 
# ggsave(
#   filename = "figure/05-figure06_1.tiff",
#   plot = figure06_1,
#   height = 5.5, width = 7
# )
```

```{r,fig.height=5.5,fig.width=7}
figure06_2 <- 
  data_output_04 %>% 
  ggplot() +
  geom_line(aes(x = estimate,y = comp.1), colour="red", lwd = 1.5) +
  geom_line(aes(x = estimate,y = comp.23), colour="blue", lwd = 1.5) +
  geom_hline(yintercept = class_probability, col = "black") +
  facet_wrap(~experimento+component_n,scales = "free",labeller = hospital_labeller) +
  labs(
    # title="Linear distribution",
    x = "Arbitrary Antibody Units",
    y = "Classification Probability",
    caption = "red : seronegative\nblue : seropositive\nblack : 90% classification probability"
  ) +
  theme(plot.caption = element_text(hjust = 0)) # set the left align here

figure06_2

# ggsave(
#   filename = "figure/05-figure06_2.png",
#   plot = figure06_2,
#   height = 5.5, width = 7
# )
# 
# ggsave(
#   filename = "figure/05-figure06_2.tiff",
#   plot = figure06_2,
#   height = 5.5, width = 7
# )
```

#### choice

```{r,fig.height=5.5,fig.width=7}
figure06 <- 
  data_output_04 %>% 
  filter(component_n == 2) %>% 
  ggplot() +
  geom_line(aes(x = estimate,y = comp.1), colour="red", lwd = 1.5) +
  geom_line(aes(x = estimate,y = comp.23), colour="blue", lwd = 1.5) +
  geom_hline(yintercept = class_probability, col = "black") +
  facet_wrap(~experimento#+component_n
             ,scales = "free",labeller = hospital_labeller) +
  labs(
    # title="Linear distribution",
    x = "Arbitrary Antibody Units",
    y = "Classification Probability",
    caption = "Red: seronegative\nBlue: seropositive\nBlack: 90% probability"
  ) +
  theme(plot.caption = element_text(hjust = 0)) # set the left align here

figure06

ggsave(
  filename = "figure/05-figure06-comp2.png",
  plot = figure06,
  height = 3.5, width = 7
)

ggsave(
  filename = "figure/05-figure06-comp2.tiff",
  plot = figure06,
  height = 3.5, width = 7
)
```


## select model 

First explore equivalent class labelling

```{r}
data_output_04 %>% 
  filter(experimento == "igg" & component_n == 2 |
           experimento == "igg-avidez" & component_n == 2) %>% 
  count(experimento,class,class_sum)
```

Then, select model by `k-component` and `k-labeling` column

```{r}
data_output_05 <- data_output_04 %>% 
  filter(experimento == "igg" & component_n == 2 |
           experimento == "igg-avidez" & component_n == 2) %>% 
  select(-class_sum)

data_output_05 %>% 
  count(experimento,component_n)
```

```{r}
data_output_05 %>% rmarkdown::paged_table()
```

# outputs

```{r}
data_output_05 %>% 
  write_rds("data/05-data-clasification-ab_units.rds")

data_output_05 %>% 
  writexl::write_xlsx("data/05-data-clasification-ab_units.xlsx")
```

# open questions

- [ ] how to deal with uncertainty in the estimate and classification (which is only based on the point estimate currently)
- [ ] how to deal with missing values? current status: excluded from classification algorithm and final data output

# to do list

- [x] clean output data table number in components
- [x] export final table to user
- [x] create unique with uncommon (?) hash `--`
- [x] select one `class` column 
- [ ] compare traditional vs data-base classification (new file)
- [ ] add workflow to https://app.diagrams.net/#G1WQ_PMmH3x6JMVnYcSdkagbiA_kiOyBG3

later

- [ ] goal: replace `stat_mixture` with sth that add the color labels in legends
- [ ] replace `plot_mix_comps` with broom like function for tidy tibble model summary
- [ ] replace density plots by automatic color based plots
- [ ] remove redundant n=3 plots
- ref: https://www.hindawi.com/journals/jir/2015/738030/
- ref: https://github.com/avallecam/elixr/blob/master/00-appendix.Rmd
- [ ] replace this 05-files workflow with the most simple univariate example --> data from https://malariajournal.biomedcentral.com/articles/10.1186/s12936-018-2471-6
