==> 06-compare_classification.Rmd <==
---
title: "06-compare_classification"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# packages

```{r,echo=TRUE,warning=FALSE,message=FALSE}
library(tidyverse)
library(janitor)
library(conflicted)
conflicts_prefer(dplyr::select)
conflicts_prefer(dplyr::filter)
```

```{r}
options(scipen=10000)
```

# import input data

```{r}
data_output_03 <- read_rds("data/03-data-ab_units-complete.rds")
data_output_05 <- read_rds("data/05-data-clasification-ab_units.rds")
```

```{r}
data_output_06 <- data_output_03 %>%
  dplyr::select(plate,experimento,estimation) %>% 
  unnest(cols = estimation) %>% 
  
  # extra code to create unique id
  mutate(unique_id = paste0(#experimento,"-",
                            plate,"--",codigo)) %>% 
  left_join(
    data_output_05 %>% 
      select(experimento,unique_id,class)
  ) #%>% 
  # mutate(
    # class = fct_na_value_to_level(class,level = "s0")
  # )
```

```{r,eval=FALSE}
# quality control
data_output_05 %>% pull(class) %>% levels()
data_output_06 %>% pull(class) %>% levels()
data_output_06 %>% count(class)
```

```{r}
data_output_06 %>% rmarkdown::paged_table()
```

# use conventional classification

```{r}
data_output_07 <- data_output_06 %>% 
  # only by experiment, assuming comparability across plates
  group_by(experimento) %>% 
  nest() %>% 
  ungroup() %>% 
  # task: summarize the population of negatives per experiment
  # [not applied] alternative: treat all ctrl_negative as units, not as replicates
  mutate(only_negatives = map(data,filter,codigo == "ctrl_negativo")) %>% 
  # od
  mutate(od_negatives = map(only_negatives,pull,densidad_mean)) %>% 
  mutate(od_cutoff_mean = map_dbl(od_negatives,mean,na.rm = T)) %>% 
  mutate(od_cutoff_sd = map_dbl(od_negatives,sd,na.rm = T)) %>% 
  mutate(od_cutoff_3sd = 3*od_cutoff_sd) %>% 
  # ab
  mutate(au_negatives = map(only_negatives,pull,estimate)) %>% 
  mutate(au_cutoff_mean = map_dbl(au_negatives,mean,na.rm = T)) %>% 
  mutate(au_cutoff_sd = map_dbl(au_negatives,sd,na.rm = T)) %>% 
  mutate(au_cutoff_3sd = 3*au_cutoff_sd) %>%
  # end
  group_by(experimento) %>% 
  mutate(cutoff_conventional_od = sum(c_across(cols = c(od_cutoff_mean,od_cutoff_3sd)),na.rm = T)) %>% 
  mutate(cutoff_conventional_au = sum(c_across(cols = c(au_cutoff_mean,au_cutoff_3sd)),na.rm = T)) %>% 
  ungroup() %>%
  select(-only_negatives,-od_negatives,-au_negatives) %>% 
  unnest(cols = c(data)) %>%
  
  # # quality control
  # count(codigo)
  # identity()
  
  mutate(class_conventional_od = case_when(
    densidad_mean <= cutoff_conventional_od ~ "s-",
    densidad_mean > cutoff_conventional_od ~ "s+",
    TRUE ~ "s0"
  )) %>% 
  mutate(class_conventional_au = case_when(
    estimate <= cutoff_conventional_au ~ "s-",
    estimate > cutoff_conventional_au ~ "s+",
    TRUE ~ "s0"
  ))
```

```{r}
data_output_07 %>% rmarkdown::paged_table()
```


# explore comparison of classification

## explore comparison 1

- Gold standard: GMM classification using standardized AU
- Comparison: Conventional classification using un-standardized OD

```{r}
data_output_07 %>%
  count(experimento,class,class_conventional_od)
```

```{r}
data_output_07 %>% 
  tabyl(class_conventional_od,class,experimento)
```

### igg

- make careful interpretations of results bellow

```{r}
data_output_07 %>% 
  filter(experimento == "igg") %>% 
  filter(class != "s0") %>% 
  filter(class != "s++") %>% 
  select(class_conventional_od,class) %>% 
  mutate(across(.cols = everything(),.fns = as.character)) %>% 
  table() %>% 
  caret::confusionMatrix("s+")
```

### igg-avidez

- make careful interpretations of results bellow

```{r}
data_output_07 %>% 
  filter(experimento == "igg-avidez") %>% 
  filter(class != "s0") %>% 
  filter(class != "s++") %>% 
  select(class_conventional_od,class) %>% 
  mutate(across(.cols = everything(),.fns = as.character)) %>% 
  table() %>% 
  caret::confusionMatrix("s+")
```

<!--

## explore comparison 2

- Gold standard: GMM classification using standardized AU
- Comparison: Conventional classification using standardized AU

Here I drop the observations with `NaN` in AU estimation

```{r}
data_output_07 %>%
  filter(!class_conventional_au=="s0") %>% 
  count(experimento,class,class_conventional_au)
```

```{r}
data_output_07 %>% 
  filter(!class_conventional_au=="s0") %>% 
  tabyl(class_conventional_au,class,experimento)
```

## explore comparison 3

- Gold standard: Conventional classification using standardized AU
- Comparison: Conventional classification using un-standardized OD

Here I drop the observations with `NaN` in AU estimation

```{r}
data_output_07 %>%
  # filter(!class_conventional_au=="s0") %>% 
  count(experimento,class_conventional_od,class_conventional_au)
```

```{r}
data_output_07 %>% 
  # filter(!class_conventional_au=="s0") %>% 
  tabyl(class_conventional_od,class_conventional_au,experimento)
```

-->

## explore data

```{r}
data_output_07 %>% 
  filter(class_conventional_au == "s0" &
           (class_conventional_od == "s+" | class_conventional_od == "s-")) %>% 
  rmarkdown::paged_table()
```


<!-- # interactive table -->

<!-- ```{r} -->
<!-- DT::datatable(data_output_07) -->
<!-- ``` -->

# outputs

```{r}
data_output_07 %>% 
  write_rds("data/07-data-clasification-compare.rds")

data_output_07 %>% 
  writexl::write_xlsx("data/07-data-clasification-compare.xlsx")
```

# to do list

- [x] move class `NA` to s0
- [x] Try: Using standardized AU to create `ctrl_negative + 3*sd` conventional cutoff?
- [ ] interactive table with filters within columns?
